Improve aptly subprocess behavior and error handling.
Aptly is apt to interactively prompt for a GPG passphrase. By using
system() instead of backticks, Ruby will present that prompt to the
user. There's also a ton of extra aptly output to stderr that helps show
what is happening.

Aptly subprocesses used to fail silently and permit Nemesis-Ops to
continue, with a corrupt repository in S3. Failures in Aptly now turn
into exceptions. Since there's no reason to expect that a nemesis
packaging workflow will be recoverable if a call to aptly fails, exiting
the app with an exception is the correct behavior.