Update the SHA for stankevich/puppet-python 1.9.1, which changed. (?!)
The old sha cf744a412... is no longer to be found in a checkout of
stankevich/puppet-python, and the release tag now points to a different
SHA. So I guess the upstream project rewrote the history of a release
tag????? That's not good.

In this case, I propose to cross my fingers and hope that we're still
importing the same code today that we were importing yesterday.