Fork and pin the Aurora package build dependencies.
The proximate cause of this patch is that the latest version of
aurora-packaging wants to use Gradle 2.7, which is not compatible with
our pinned version of Aurora itself.

At the same time, I'm proposing an alternative means of checking out
Acquia Github repositories for the Aurora build, using Github OAuth
tokens. We can then:

  - Create a Grid-Jenkins-only Github user that has read-only access to
    Grid codebases.

  - Create multiple Github tokens for that user and put those tokens on
    Jenkins instances in prod, integration, or dev Grids.

  - Independently create and revoke those Github tokens using Github's
    UI or API.